+++ frontend/src/features/auth/components/Auth/ProtectedRoute.tsx
@@ -1,0 +1,18 @@
+import useAuth from '../../hooks/useAuth';
+import { Navigate, Outlet } from 'react-router-dom';
+
+const ProtectedRoute = () => {
+  const { isAuthenticated, loading } = useAuth();
+
+  if (loading) {
+    return <div>Loading...</div>;
+  }
+
+  if (!isAuthenticated) {
+    return <Navigate to='/' replace/>;
+  }
+
+  return <Outlet />
+};
+
+export default ProtectedRoute;
