--- frontend/src/features/auth/components/ProtectedRoute.tsx
@@ -1,23 +1,0 @@
-import React from 'react';
-import useAuth from '../hooks/useAuth';
-import { Navigate } from 'react-router-dom';
-
-interface ProtectedRouteProps {
-  children: React.ReactNode;
-}
-
-const ProtectedRoute = ({ children }: ProtectedRouteProps) => {
-  const { isAuthenticated, loading } = useAuth();
-
-  if (loading) {
-    return <div>Loading...</div>;
-  }
-
-  if (!isAuthenticated) {
-    return <Navigate to={'/'} />;
-  }
-
-  return <>{children}</>;
-};
-
-export default ProtectedRoute;
